convplot

Note: slightly more up to date forks in convplot-dev repository https://github.com/mdingemanse/convoplot-dev

Usage

convplot(
  data,
  uids = NULL,
  lang = NULL,
  n_uid = 10,
  window = NULL,
  before = 10000,
  after = 10000,
  printuids = T,
  verbose = T,
  dyads = F,
  content = F,
  highlight = F,
  center = F,
  datamode = F,
  alldata = F,
  debug = F
)

Arguments

data

dataset

uids

set of uids to plot (optional; if omitted, n uids are sampled)

lang

language from which to sample uids (if not supplied)

n_uid

number of uids to sample

window

time window in ms (optional; if supplied, window will be split into before and after)

before

stretch to include before selected turn (default: 10000ms, unless `window` is supplied)

after

stretch to include after selected turn (default: 0, unless `window` is supplied)

printuids

print the sampled uids

verbose

print language and information about selected uids

dyads

if TRUE, select only dyadic interactions for display

content

if TRUE, render annotation content (EXPERIMENTAL)

highlight

if TRUE, highlight the uid in focus in the plot

center

if TRUE, center the plot around the uid in focus

datamode

if TRUE, outputs dataframe instead of plot, for more advanced plotting

alldata

if TRUE, output all data, not just the selected dyads

debug

if TRUE, print the selected data and some other diagnostics